pyproject.toml
[tool.poetry]
name = "model-runner-client"
version = "0.9.0"
description = "Model Runner Client is a Python library that allows Coordinators to manage real-time model synchronization and perform concurrent predictions on distributed model nodes within a Crunch."
authors = ["boutrig abdennour <abdennour.boutrig@crunchdao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1"
pyarrow = { version = "^22.0.0", optional = true }
pandas = "^2.2.3"
websockets = "^14.2"
grpcio-health-checking = "^1.74.0"
cryptography = "^46.0.3"
base58 = "^2.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[project.urls]
#"Homepage" = "https://github.com/crunchdao/model-runner-client"
#"Bug Tracker" = "https://github.com/crunchdao/model-runner-client/issues"
